+++ backend/src/main/java/com/blsa/ezilog/service/point/PointService.java
@@ -10,13 +10,13 @@ import com.blsa.ezilog.model.point.PointHistory;
 
 public interface PointService {
      
-    public Page<PointHistory> selectPointByUser(BigInteger Uid, Pageable pageable);
+    public Page<PointHistory> selectPointByUser(Long Uid, Pageable pageable);
     
     List<PointHistory> selectRankByTotalPoint();
     
     boolean addPoint(PointHistory point);
     
-    int totalPoint(BigInteger Uid);
+    int totalPoint(Long Uid);
     
-    void deletePoint(BigInteger id);
+    void deletePoint(Long id);
 }
